Refactor: break large modules into smaller, focused files
Extract six concerns from the three largest files (tasks.py, app.py, cli/main.py):

- lib/terminal.py: consolidated color/terminal helpers (from tasks.py + cli/main.py)
- lib/clipboard.py: clipboard utilities (from tasks.py)
- lib/containers.py: podman container management (from tasks.py + widgets.py)
- tui/screens.py: ProjectActionsScreen + TaskActionsScreen modals (from app.py)
- tui/polling.py: PollingMixin for upstream/container polling (from app.py)
- cli/main.py: extract _print_config() from monolithic main()

Updates test mock paths to target new canonical module locations.
