+++ src/luskctl/tui/app.py
@@ -30,22 +30,17 @@ except Exception:  # pragma: no cover - textual not installed
 
 if _HAS_TEXTUAL:
     # Import textual and our widgets only when available
-    from textual import events, on, screen
+    from textual import on
     from textual.app import App, ComposeResult
-
-    try:  # pragma: no cover - optional import for test stubs
-        from textual.binding import Binding
-    except Exception:  # pragma: no cover - textual may be a stub module
-        Binding = None  # type: ignore[assignment]
     from textual.containers import Horizontal, Vertical
     from textual.widgets import Button, Footer, Header
     from textual.worker import Worker, WorkerState
 
+    from ..lib.clipboard import copy_to_clipboard_detailed, get_clipboard_helper_status
+    from ..lib.containers import _is_task_image_old
     from ..lib.docker import build_images, generate_dockerfiles
     from ..lib.git_gate import (
         GateStalenessInfo,
-        compare_gate_vs_upstream,
-        sync_gate_branches,
         sync_project_gate,
     )
     from ..lib.projects import Project as CodexProject
@@ -53,8 +48,6 @@ if _HAS_TEXTUAL:
     from ..lib.ssh import init_project_ssh
     from ..lib.tasks import (
         WEB_BACKENDS,
-        copy_to_clipboard_detailed,
-        get_clipboard_helper_status,
         get_tasks,
         get_workspace_git_diff,
         task_delete,
@@ -65,6 +58,8 @@ if _HAS_TEXTUAL:
 
     # Import version info function from lib (shared with CLI --version)
     from ..lib.version import get_version_info as _get_version_info
+    from .polling import PollingMixin
+    from .screens import ProjectActionsScreen, TaskActionsScreen
     from .widgets import (
         ProjectList,
         ProjectState,
@@ -72,378 +67,9 @@ if _HAS_TEXTUAL:
         TaskDetails,
         TaskList,
         TaskMeta,
-        _is_task_image_old,
     )
 
-    def _modal_binding(key: str, action: str, description: str):
-        if Binding is None:
-            return (key, action, description)
-        return Binding(key, action, description, show=False)
-
-    class ProjectActionsScreen(screen.ModalScreen[str | None]):
-        """Modal screen for project actions."""
-
-        BINDINGS = [
-            _modal_binding("escape", "dismiss", "Cancel"),
-            _modal_binding("q", "dismiss", "Cancel"),
-            _modal_binding("up", "app.focus_previous", "Previous"),
-            _modal_binding("down", "app.focus_next", "Next"),
-            _modal_binding("d", "generate", "Generate dockerfiles"),
-            _modal_binding("b", "build", "Build project image"),
-            _modal_binding("a", "build_agents", "Rebuild with fresh agents"),
-            _modal_binding("f", "build_full", "Full rebuild no cache"),
-            _modal_binding("s", "init_ssh", "Init SSH"),
-            _modal_binding("g", "sync_gate", "Sync git gate"),
-        ]
-
-        COMPACT_HEIGHT = 20
-
-        CSS = """
-        ProjectActionsScreen {
-            align: center middle;
-            padding: 1 0;
-        }
-
-        #action-dialog {
-            width: 60;
-            height: auto;
-            max-height: 100%;
-            border: heavy $primary;
-            border-title-align: right;
-            background: $surface;
-            padding: 1 1 0 1;
-            overflow-y: auto;
-        }
-
-        #action-buttons {
-            layout: vertical;
-            margin-top: 1;
-            align: center middle;
-            width: 100%;
-        }
-
-        #action-cancel {
-            margin-top: 0;
-            align: center middle;
-            width: 100%;
-        }
-
-        #action-buttons Button {
-            margin: 0 0 1 0;
-            width: 100%;
-            min-width: 0;
-        }
-
-        #action-dialog Button {
-            width: 100%;
-            min-width: 0;
-        }
-
-        ProjectActionsScreen.compact #action-dialog {
-            padding: 0 1;
-        }
-
-        ProjectActionsScreen.compact #action-buttons {
-            margin-top: 0;
-        }
-
-        ProjectActionsScreen.compact #action-dialog Button {
-            border: none;
-            height: 1;
-            padding: 0 1;
-        }
-
-        """
-
-        def __init__(self, title: str | None = None) -> None:
-            super().__init__()
-            self._dialog_title = title or "Project Actions"
-
-        def compose(self) -> ComposeResult:
-            with Vertical(id="action-dialog") as dialog:
-                with Vertical(id="action-buttons"):
-                    yield Button(
-                        "generate [yellow]d[/yellow]ockerfiles",
-                        id="generate",
-                        variant="primary",
-                    )
-                    yield Button(
-                        "[yellow]b[/yellow]uild project image",
-                        id="build",
-                        variant="primary",
-                    )
-                    yield Button(
-                        "rebuild with [yellow]a[/yellow]gents",
-                        id="build_agents",
-                        variant="primary",
-                    )
-                    yield Button(
-                        "[yellow]f[/yellow]ull rebuild no cache",
-                        id="build_full",
-                        variant="primary",
-                    )
-                    yield Button(
-                        "initialize [yellow]s[/yellow]sh", id="init_ssh", variant="primary"
-                    )
-                    yield Button(
-                        "sync [yellow]g[/yellow]it gate",
-                        id="sync_gate",
-                        variant="primary",
-                    )
-                with Horizontal(id="action-cancel"):
-                    yield Button("Cancel", id="cancel", variant="default")
-            dialog.border_title = self._dialog_title
-
-        def on_button_pressed(self, event: Button.Pressed) -> None:
-            button_id = event.button.id
-            if button_id == "cancel":
-                self.dismiss(None)
-            else:
-                action_map = {
-                    "generate": "generate",
-                    "build": "build",
-                    "build_agents": "build_agents",
-                    "build_full": "build_full",
-                    "sync_gate": "sync_gate",
-                    "init_ssh": "init_ssh",
-                }
-                self.dismiss(action_map.get(button_id))
-
-        def on_key(self, event: events.Key) -> None:
-            key = event.key.lower()
-            if key in {"escape", "q"}:
-                self.action_dismiss()
-                event.stop()
-                return
-            if key == "up":
-                if self.app:
-                    self.app.action_focus_previous()
-                event.stop()
-                return
-            if key == "down":
-                if self.app:
-                    self.app.action_focus_next()
-                event.stop()
-                return
-            if key == "d":
-                self.action_generate()
-                event.stop()
-                return
-            if key == "b":
-                self.action_build()
-                event.stop()
-                return
-            if key == "a":
-                self.action_build_agents()
-                event.stop()
-                return
-            if key == "f":
-                self.action_build_full()
-                event.stop()
-                return
-            if key == "s":
-                self.action_init_ssh()
-                event.stop()
-                return
-            if key == "g":
-                self.action_sync_gate()
-                event.stop()
-
-        def action_dismiss(self) -> None:
-            self.dismiss(None)
-
-        def action_generate(self) -> None:
-            self.dismiss("generate")
-
-        def action_build(self) -> None:
-            self.dismiss("build")
-
-        def action_build_agents(self) -> None:
-            self.dismiss("build_agents")
-
-        def action_build_full(self) -> None:
-            self.dismiss("build_full")
-
-        def action_init_ssh(self) -> None:
-            self.dismiss("init_ssh")
-
-        def action_sync_gate(self) -> None:
-            self.dismiss("sync_gate")
-
-        def on_mount(self) -> None:
-            first_button = self.query_one("#action-buttons Button", Button)
-            first_button.focus()
-            self._update_density()
-
-        def on_resize(self, _event: object) -> None:
-            self._update_density()
-
-        def _update_density(self) -> None:
-            self.set_class(self.size.height < self.COMPACT_HEIGHT, "compact")
-
-    class TaskActionsScreen(screen.ModalScreen[str | None]):
-        """Modal screen for task actions."""
-
-        BINDINGS = [
-            _modal_binding("escape", "dismiss", "Cancel"),
-            _modal_binding("q", "dismiss", "Cancel"),
-            _modal_binding("up", "app.focus_previous", "Previous"),
-            _modal_binding("down", "app.focus_next", "Next"),
-            _modal_binding("n", "new_task", "New task"),
-            _modal_binding("c", "cli", "CLI agent"),
-            _modal_binding("w", "web", "Web UI"),
-            _modal_binding("d", "delete", "Delete task"),
-        ]
-
-        COMPACT_HEIGHT = 18
-
-        CSS = """
-        TaskActionsScreen {
-            align: center middle;
-            padding: 1 0;
-        }
-
-        #action-dialog {
-            width: 60;
-            height: auto;
-            max-height: 100%;
-            border: heavy $primary;
-            border-title-align: right;
-            background: $surface;
-            padding: 1 1 0 1;
-            overflow-y: auto;
-        }
-
-        #action-buttons {
-            layout: vertical;
-            margin-top: 1;
-            align: center middle;
-            width: 100%;
-        }
-
-        #action-cancel {
-            margin-top: 0;
-            align: center middle;
-            width: 100%;
-        }
-
-        #action-buttons Button {
-            margin: 0 0 1 0;
-            width: 100%;
-            min-width: 0;
-        }
-
-        #action-dialog Button {
-            width: 100%;
-            min-width: 0;
-        }
-
-        TaskActionsScreen.compact #action-dialog {
-            padding: 0 1;
-        }
-
-        TaskActionsScreen.compact #action-buttons {
-            margin-top: 0;
-        }
-
-        TaskActionsScreen.compact #action-dialog Button {
-            border: none;
-            height: 1;
-            padding: 0 1;
-        }
-
-        """
-
-        def __init__(self, title: str | None = None, *, has_tasks: bool = True) -> None:
-            super().__init__()
-            self._dialog_title = title or "Task Actions"
-            self._has_tasks = has_tasks
-
-        def compose(self) -> ComposeResult:
-            with Vertical(id="action-dialog") as dialog:
-                with Vertical(id="action-buttons"):
-                    yield Button("[yellow]n[/yellow]ew task", id="new", variant="primary")
-                    if self._has_tasks:
-                        yield Button("[yellow]c[/yellow]li agent", id="cli", variant="primary")
-                        yield Button("[yellow]w[/yellow]eb ui", id="web", variant="primary")
-                        yield Button("[yellow]d[/yellow]elete task", id="delete", variant="primary")
-                with Horizontal(id="action-cancel"):
-                    yield Button("Cancel", id="cancel", variant="default")
-            dialog.border_title = self._dialog_title
-
-        def on_button_pressed(self, event: Button.Pressed) -> None:
-            button_id = event.button.id
-            if button_id == "cancel":
-                self.dismiss(None)
-            else:
-                action_map = {"new": "new", "cli": "cli", "web": "web", "delete": "delete"}
-                self.dismiss(action_map.get(button_id))
-
-        def on_key(self, event: events.Key) -> None:
-            key = event.key.lower()
-            if key in {"escape", "q"}:
-                self.action_dismiss()
-                event.stop()
-                return
-            if key == "up":
-                if self.app:
-                    self.app.action_focus_previous()
-                event.stop()
-                return
-            if key == "down":
-                if self.app:
-                    self.app.action_focus_next()
-                event.stop()
-                return
-            if key == "n":
-                self.action_new_task()
-                event.stop()
-                return
-            if key == "c":
-                self.action_cli()
-                event.stop()
-                return
-            if key == "w":
-                self.action_web()
-                event.stop()
-                return
-            if key == "d":
-                self.action_delete()
-                event.stop()
-
-        def action_dismiss(self) -> None:
-            self.dismiss(None)
-
-        def action_new_task(self) -> None:
-            self.dismiss("new")
-
-        def action_cli(self) -> None:
-            if not self._has_tasks:
-                return
-            self.dismiss("cli")
-
-        def action_web(self) -> None:
-            if not self._has_tasks:
-                return
-            self.dismiss("web")
-
-        def action_delete(self) -> None:
-            if not self._has_tasks:
-                return
-            self.dismiss("delete")
-
-        def on_mount(self) -> None:
-            first_button = self.query_one("#action-buttons Button", Button)
-            first_button.focus()
-            self._update_density()
-
-        def on_resize(self, _event: object) -> None:
-            self._update_density()
-
-        def _update_density(self) -> None:
-            self.set_class(self.size.height < self.COMPACT_HEIGHT, "compact")
-
-    class LuskTUI(App):
+    class LuskTUI(PollingMixin, App):
         """Redesigned TUI frontend for luskctl core modules."""
 
         CSS_PATH = None
@@ -937,251 +563,6 @@ if _HAS_TEXTUAL:
                 return project_id, task_id, None
             except Exception as e:
                 return project_id, task_id, str(e)
-
-        # ---------- Upstream polling ----------
-
-        def _start_upstream_polling(self) -> None:
-            """Start background polling for upstream changes.
-
-            Only polls for gatekeeping projects with polling enabled and a gate initialized.
-            """
-            self._stop_upstream_polling()  # Stop any existing timer
-            self._staleness_info = None
-            self._last_notified_stale = False
-
-            if not self.current_project_id:
-                return
-
-            try:
-                project = load_project(self.current_project_id)
-            except SystemExit:
-                return
-
-            # Only poll for gatekeeping projects with polling enabled
-            if project.security_class != "gatekeeping":
-                return
-            if not project.upstream_polling_enabled:
-                return
-            if not project.gate_path.exists():
-                return
-
-            interval_seconds = project.upstream_polling_interval_minutes * 60
-            self._polling_project_id = self.current_project_id
-
-            # Perform initial poll immediately (in background worker)
-            self._poll_upstream()
-
-            # Schedule recurring polls
-            self._polling_timer = self.set_interval(
-                interval_seconds, self._poll_upstream, name="upstream_polling"
-            )
-
-        def _stop_upstream_polling(self) -> None:
-            """Stop the upstream polling timer."""
-            if self._polling_timer is not None:
-                self._polling_timer.stop()
-                self._polling_timer = None
-            self._polling_project_id = None
-
-        def _start_container_status_polling(self) -> None:
-            """Start background polling for container status every 2 seconds."""
-            self._stop_container_status_polling()
-            if not self.current_project_id:
-                return
-            # Poll every 2 seconds - podman inspect is fast (~10-50ms)
-            interval_seconds = 2
-            # Initial poll
-            self._poll_container_status()
-            # Schedule recurring polls
-            self._container_status_timer = self.set_interval(
-                interval_seconds, self._poll_container_status, name="container_status_polling"
-            )
-
-        def _stop_container_status_polling(self) -> None:
-            """Stop the container status polling timer."""
-            if self._container_status_timer is not None:
-                self._container_status_timer.stop()
-                self._container_status_timer = None
-
-        def _poll_container_status(self) -> None:
-            """Check container status for the current task."""
-            if not self.current_project_id or not self.current_task or not self.current_task.mode:
-                return
-            self._queue_container_state_check(
-                self.current_project_id,
-                self.current_task.task_id,
-                self.current_task.mode,
-            )
-
-        def _queue_container_state_check(self, project_id: str, task_id: str, mode: str) -> None:
-            """Queue a background check for a task's container state."""
-            self.run_worker(
-                self._load_container_state_worker(project_id, task_id, mode),
-                name=f"container-state:{project_id}:{task_id}",
-                group="container-state",
-                exclusive=True,
-            )
-
-        async def _load_container_state_worker(
-            self, project_id: str, task_id: str, mode: str
-        ) -> tuple[str, str, str | None]:
-            """Background worker to check container state."""
-            import asyncio
-
-            from ..lib.tasks import get_task_container_state
-
-            try:
-                state = await asyncio.get_event_loop().run_in_executor(
-                    None, get_task_container_state, project_id, task_id, mode
-                )
-                return (project_id, task_id, state)
-            except Exception as e:
-                self._log_debug(f"container state check error: {e}")
-                return (project_id, task_id, None)
-
-        def _poll_upstream(self) -> None:
-            """Check upstream for changes and update staleness info.
-
-            Runs the actual comparison in a background worker to avoid blocking the UI.
-            """
-            project_id = self._polling_project_id
-            if not project_id or project_id != self.current_project_id:
-                # Project changed since timer was started, skip this poll
-                return
-
-            self._log_debug(f"polling upstream for {project_id}")
-            # Run blocking git operation in background worker
-            self.run_worker(
-                self._poll_upstream_worker(project_id),
-                name="poll_upstream",
-                exclusive=True,  # Cancel any previous poll still running
-            )
-
-        async def _poll_upstream_worker(self, project_id: str) -> None:
-            """Background worker to check upstream (runs in thread pool)."""
-            import asyncio
-
-            try:
-                # Run blocking call in thread pool
-                staleness = await asyncio.get_event_loop().run_in_executor(
-                    None, compare_gate_vs_upstream, project_id
-                )
-
-                # Validate project hasn't changed while we were polling
-                if project_id != self.current_project_id:
-                    return
-
-                self._on_staleness_updated(project_id, staleness)
-
-            except Exception as e:
-                self._log_debug(f"upstream poll error: {e}")
-
-        def _on_staleness_updated(self, project_id: str, staleness: GateStalenessInfo) -> None:
-            """Handle updated staleness info."""
-            # Double-check project hasn't changed
-            if project_id != self.current_project_id:
-                return
-
-            self._staleness_info = staleness
-
-            # Only update notification state for valid (non-error) comparisons
-            if staleness.error:
-                # Don't change notification state on errors - preserve previous state
-                pass
-            elif staleness.is_stale and not self._last_notified_stale:
-                behind_str = ""
-                if staleness.commits_behind is not None:
-                    behind_str = f" ({staleness.commits_behind} commits behind)"
-                self.notify(f"Gate is behind upstream on {staleness.branch}{behind_str}")
-                self._last_notified_stale = True
-
-                # Trigger auto-sync if enabled (with cooldown check)
-                self._maybe_auto_sync(project_id)
-            elif not staleness.is_stale:
-                # Only reset when we have confirmed up-to-date status
-                self._last_notified_stale = False
-
-            # Refresh the project state display
-            self._refresh_project_state()
-
-        def _maybe_auto_sync(self, project_id: str) -> None:
-            """Trigger auto-sync if enabled for this project.
-
-            Runs sync in background worker to avoid blocking UI.
-            Implements cooldown to prevent sync loops.
-            """
-            import time
-
-            if not project_id or project_id != self.current_project_id:
-                return
-
-            # Check cooldown (5 minute minimum between auto-syncs per project)
-            now = time.time()
-            cooldown_until = self._auto_sync_cooldown.get(project_id, 0)
-            if now < cooldown_until:
-                self._log_debug("auto-sync skipped: cooldown active")
-                return
-
-            try:
-                project = load_project(project_id)
-                if not project.auto_sync_enabled:
-                    return
-
-                # Set cooldown before starting sync (5 minutes)
-                self._auto_sync_cooldown[project_id] = now + 300
-
-                self._log_debug(f"auto-syncing gate for {project_id}")
-                self.notify("Auto-syncing gate from upstream...")
-
-                # Run sync in background worker
-                branches = project.auto_sync_branches or None
-                self.run_worker(
-                    self._sync_worker(project_id, branches, is_auto=True),
-                    name="auto_sync",
-                    exclusive=True,
-                )
-
-            except Exception as e:
-                self._log_debug(f"auto-sync error: {e}")
-
-        async def _sync_worker(
-            self, project_id: str, branches: list = None, is_auto: bool = False
-        ) -> None:
-            """Background worker to sync gate from upstream."""
-            import asyncio
-
-            try:
-                # Run blocking sync in thread pool
-                result = await asyncio.get_event_loop().run_in_executor(
-                    None, sync_gate_branches, project_id, branches
-                )
-
-                # Validate project hasn't changed
-                if project_id != self.current_project_id:
-                    return
-
-                if result["success"]:
-                    label = "Auto-synced" if is_auto else "Synced"
-                    self.notify(f"{label} gate from upstream")
-
-                    # Re-check staleness after sync
-                    staleness = await asyncio.get_event_loop().run_in_executor(
-                        None, compare_gate_vs_upstream, project_id
-                    )
-
-                    if project_id == self.current_project_id:
-                        self._staleness_info = staleness
-                        # Only reset notification flag if we're actually up-to-date now
-                        if not staleness.is_stale and not staleness.error:
-                            self._last_notified_stale = False
-                        self._refresh_project_state()
-                else:
-                    label = "Auto-sync" if is_auto else "Sync"
-                    self.notify(f"{label} failed: {', '.join(result['errors'])}")
-
-            except Exception as e:
-                label = "Auto-sync" if is_auto else "Sync"
-                self.notify(f"{label} error: {e}")
 
         async def action_sync_gate(self) -> None:
             """Manually sync gate from upstream."""
