+++ src/luskctl/tui/widgets.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 
 import inspect
-import subprocess
 from dataclasses import dataclass
 from typing import Any
 
@@ -92,71 +91,6 @@ def _get_css_variables(widget: Static) -> dict[str, str]:
         return widget.app.get_css_variables()
     except Exception:
         return {}
-
-
-def _is_task_image_old(project_id: str | None, task: TaskMeta) -> bool | None:
-    if project_id is None:
-        return None
-    if task.mode not in {"cli", "web"}:
-        return None
-
-    container_name = f"{project_id}-{task.mode}-{task.task_id}"
-    try:
-        result = subprocess.run(
-            [
-                "podman",
-                "container",
-                "inspect",
-                "--format",
-                "{{.State.Running}}\t{{.Image}}",
-                container_name,
-            ],
-            capture_output=True,
-            text=True,
-            timeout=10,
-        )
-    except (FileNotFoundError, OSError, subprocess.TimeoutExpired):
-        return None
-    if result.returncode != 0:
-        return None
-
-    running_str, _, image_id = result.stdout.partition("\t")
-    if running_str.strip().lower() != "true":
-        return None
-    image_id = image_id.strip()
-    if not image_id:
-        return None
-
-    try:
-        from ..lib.docker import build_context_hash
-
-        current_hash = build_context_hash(project_id)
-    except Exception:
-        return None
-
-    try:
-        label_result = subprocess.run(
-            [
-                "podman",
-                "image",
-                "inspect",
-                "--format",
-                '{{index .Config.Labels "luskctl.build_context_hash"}}',
-                image_id,
-            ],
-            capture_output=True,
-            text=True,
-            timeout=10,
-        )
-    except (FileNotFoundError, OSError, subprocess.TimeoutExpired):
-        return None
-    if label_result.returncode != 0:
-        return None
-
-    label = label_result.stdout.strip()
-    if not label or label == "<no value>":
-        return True
-    return label != current_hash
 
 
 class ProjectList(ListView):
